# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  address?: string;
  city?: string;
  country?: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  categoryId: number;
  categoryName: string;
  /** @nullable */
  brand?: string | null;
  stock: number;
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  isDeal?: boolean;
  isTrending?: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  categoryId: number;
  brand?: string;
  stock: number;
  isFeatured?: boolean;
  isDeal?: boolean;
  isTrending?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  quantity: number;
  /** @nullable */
  imageUrl: string | null;
  stock?: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  total: number;
  shippingAddress: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  trackingNumber?: string | null;
  createdAt: string;
  /** @nullable */
  estimatedDelivery?: string | null;
}

export interface OrderInput {
  shippingAddress: string;
  city?: string;
  country?: string;
  paymentMethod: string;
  cardLast4?: string;
}

export interface OrderSummary {
  totalOrders: number;
  totalSpent: number;
  pendingOrders: number;
  deliveredOrders: number;
}

export interface Review {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  verified?: boolean;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  title?: string;
  body?: string;
}

export type GetProductsParams = {
q?: string;
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
minPrice?: number | null;
/**
 * @nullable
 */
maxPrice?: number | null;
/**
 * @nullable
 */
rating?: number | null;
sort?: GetProductsSort;
page?: number;
limit?: number;
};

export type GetProductsSort = typeof GetProductsSort[keyof typeof GetProductsSort];


export const GetProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  newest: 'newest',
  popular: 'popular',
} as const;