# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FashinHub E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # PRODUCTS
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: rating
          in: query
          schema:
            type: ["number", "null"]
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest, popular]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # CATEGORIES
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Product:
      type: object
      required: [id, name, price, categoryId, categoryName, stock, rating, reviewCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        brand:
          type: ["string", "null"]
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
        isTrending:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, categoryId, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        categoryId:
          type: integer
        brand:
          type: string
        stock:
          type: integer
        isFeatured:
          type: boolean
        isDeal:
          type: boolean
        isTrending:
          type: boolean
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Category
...
    CartItem:
      type: object
      required: [productId, productName, price, quantity, imageUrl]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
    Cart:
      type: object
      required: [items, subtotal, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        itemCount:
          type: integer
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # Orders
    OrderItem:
      type: object
      required: [productId, productName, price, quantity, imageUrl]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, status, items, subtotal, shipping, total, shippingAddress, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          type: string
        paymentMethod:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        createdAt:
          type: string
        estimatedDelivery:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [shippingAddress, paymentMethod]
      properties:
        shippingAddress:
          type: string
        city:
          type: string
        country:
          type: string
        paymentMethod:
          type: string
        cardLast4:
          type: string
...