# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  quantity: number;
  /** @nullable */
  imageUrl: string | null;
  stock?: number;
}