# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  total: number;
  shippingAddress: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  trackingNumber?: string | null;
  createdAt: string;
  /** @nullable */
  estimatedDelivery?: string | null;
}