# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  shippingAddress: string;
  city?: string;
  country?: string;
  paymentMethod: string;
  cardLast4?: string;
}