# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  categoryId: number;
  categoryName: string;
  /** @nullable */
  brand?: string | null;
  stock: number;
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  isDeal?: boolean;
  isTrending?: boolean;
  createdAt: string;
}