# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FashinHub E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  categoryId: number;
  brand?: string;
  stock: number;
  isFeatured?: boolean;
  isDeal?: boolean;
  isTrending?: boolean;
}